Moved checksums to new utility
